[REF] web: convert codeEditor tests to HOOT
task-3705027
